+++ src/tooling/piral-cli-webpack5/src/configs/piral.ts
@@ -46,8 +46,8 @@ export async function getPiralConfig(
       output: {
         publicPath,
         path: dist,
-        filename: `index.${contentHash ? '[contenthash].' : ''}js`,
-        chunkFilename: contentHash ? '[chunkhash:8].js' : undefined,
+        filename: `index.${contentHash ? '[contenthash:6].' : ''}js`,
+        chunkFilename: contentHash ? '[chunkhash:6].js' : undefined,
       },
 
       resolve: {