+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.13.3 (tbd)
 
+* Fixed missing support for `favicon-webpack-plugin` in `piral-cli-webpack`
 * Added `dashboardPath` and `piralChildren` as options for `renderInstance`
 * Updated used chunk hash length for Webpack