+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added `dashboardPath` and `piralChildren` as options for `renderInstance`
 * Updated used chunk hash length for Webpack
 * Replaced `node-sass` with `sass`
+* Renamed `master` branch to `main`
 
 ## 0.13.2 (April 27, 2021)