+++ CHANGELOG.md
@@ -171,7 +171,7 @@
 * Added support for Yarn and Pnpm (#203)
 * Added optional `--cacert` flag to `pilet publish` command (#204)
 * Fixed `NODE_ENV` mode of the emulator (#207)
-* Added `getProfile()` to [piral-oidc](./src/packages/piral-oidc/README.md) (#210)
+* Added `getProfile()` to [piral-oidc](./src/plugins/piral-oidc/README.md) (#210)
 * Switched to relative paths for local Piral references
 * Added new package `siteless` for painless microfrontends
 * Added `reducers` option to `piral-feeds`